requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1
crewai==0.11.0
openai==1.3.7
pydantic==2.5.0
python-multipart==0.0.6
requests==2.31.0
pandas==2.1.3
numpy==1.25.2
python-dotenv==1.0.0
plotly==5.17.0 